requirements.txt
Django>=4.2.8
djangorestframework>=3.14.0
pytest>=7.4.4
pytest-django>=4.10.0
Faker>=18.9.0
drf-yasg>=1.20.0
psycopg2-binary>=2.9.6
djangorestframework-simplejwt>=5.2.2